+++ themes/theme-b2b/src/components/Tag.styles.ts
@@ -6,6 +6,6 @@ export const Tag: ThemeComponent<'Tag'> = {
     'data-selected:bg-bg-selected-input data-selected:text-text-inverted data-selected:border-border-selected-input',
     'data-[disabled]:cursor-not-allowed data-[disabled]:text-text-base-disabled data-[disabled]:bg-bg-base-disabled',
   ]),
-  closeButton: cva('right-1 size-4 cursor-pointer p-0 outline-hidden'),
+  closeButton: cva('right-1 size-4'),
   listItems: cva('flex flex-wrap items-center gap-1'),
 };