+++ themes/theme-rui/src/components/Accordion.styles.ts
@@ -19,7 +19,7 @@ export const Accordion: ThemeComponent<'Accordion'> = {
         card: [
           'rounded-md border px-4 py-1 outline-none last:border-b',
           // TODO: focus is still shown even if an item within the item is focused
-          'mixin-ring-has-focus-visible',
+          'has-focus-visible:util-focus-ring',
         ],
       },
     },
@@ -37,7 +37,7 @@ export const Accordion: ThemeComponent<'Accordion'> = {
     {
       variants: {
         variant: {
-          default: 'mixin-ring-focus-visible',
+          default: 'focus-visible:util-focus-ring',
           card: 'outline-none',
         },
       },